Add files via upload
Greatly improved the way in which items will spawn, including taking into account the item quality with it affecting the players variables. The game will also not let you pickup 2 of the same item. Finally the items will be properly displayed in the players inventory with the quality of each item shown behind it.